(merge) Fix immediate pref value change in display resolution settings
We use a change observer to trigger updating the selected mode of a
display. This is useful for the drop down menu we use to select the mode
on external displays. However, if display zoom is disabled, we show a
slider instead of a drop down. The slider triggers the same pref value
change even if the user is still dragging it. This causes a usability
regression on the settings resolution slider.

To fix this, the patch adds a check which ensures we only listen to
events fired due to releasing the slider when display zoom is disabled.

Bug: 845712
Cq-Include-Trybots: master.tryserver.chromium.linux:closure_compilation
Change-Id: I9965bdf3121289ac29f38175cc9ad179cda41497
Component: Display settings, resolution slider
Reviewed-on: https://chromium-review.googlesource.com/1069773
Commit-Queue: Malay Keshav <malaykeshav@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#562555}(cherry picked from commit d88b3c4970f59187989c23885aca1b29bb5a202b [formerly 5730340bfabe9ffe98b1b72336a34ee506be4422])
Reviewed-on: https://chromium-review.googlesource.com/1099695
Reviewed-by: Malay Keshav <malaykeshav@chromium.org>
Cr-Commit-Position: refs/branch-heads/3440@{#340}
Cr-Branched-From: dc83408e893a6ef8ceae89e5588f0a97ba95eaad [formerly 010ddcfda246975d194964ccf20038ebbdec6084]-refs/heads/master@{#561733}
Former-commit-id: 638621ef4970061838486e43573acbaea12a993f