chromium: fix cursor flashes when cross between divs
On TKS ticket, cursor can flash between cross and candrop when drag
content and hover on cells. Fixing it by not apply the cursor style when
it comes to a new position at the first time. And at the next flash when
it come to drag over again. If the clientpoint is same. Then cursor will
change to the right state.

Type:
fix