Custom Tabs: Scroll publisher attribution text to end of TLD
The UrlBar used to attempt to parse the display text as a URL to figure
out the origin part of it. This fails with AMP publisher attribution,
as the display text is not a pure URL anymore. In order to avoid any
shenanigans with trying to parse the publisher attribution string,
change ToolbarDataProvider to return a UrlBarData object that contains
both the edit text and the display text, plus the start and end of the
origin in the latter. This makes it easier to avoid diverging behavior
for edit and display text, and moves the UrlBar closer to being a "dumb"
view.

Longer-term, the same mechanism could be used to identify the parts of
the URL that should be highlighted or deemphasized when returning them
from the ToolbarDataProvider, instead of UrlBar parsing the URL.

Because scrolling a TextView does not work with ellipsizing at the end,
undo the ellipsizing from https://crrev.com/549590.

(cherry picked from commit cae19600f336432a2ea9e0824add5c3af081d7dd [formerly c34f42f536db73e806108caa72f4b2fa0b7d2290])

Bug: 835317
Change-Id: Id11c5e2d0a5fb83775eed098ea3ad2be4b07bc21
Reviewed-on: https://chromium-review.googlesource.com/1017120
Commit-Queue: Bernhard Bauer <bauerb@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#563349}
Reviewed-on: https://chromium-review.googlesource.com/1101077
Reviewed-by: Bernhard Bauer <bauerb@chromium.org>
Cr-Commit-Position: refs/branch-heads/3440@{#373}
Cr-Branched-From: dc83408e893a6ef8ceae89e5588f0a97ba95eaad [formerly 010ddcfda246975d194964ccf20038ebbdec6084]-refs/heads/master@{#561733}
Former-commit-id: 04287dfb13d1fb924dfeb436aa4ab3bfe4d07d58