Composited LCD text (7/7) - blink layout changes
This commit implements the effects of the 'lcd-background-color' on the
graphics layers created by blink.

The graphics layers are created by 'blink::CompositedLayerMapping'. This
commit adds the function 'updateLCDBackgroundColor()', which will read the
'lcd-background-color' from the associated 'blink::LayoutObject'. A
'blink::CompositedLayerMapping' may be associated with a stack of
'blink::LayoutObject's, so, this function consults all such objects between
'this' 'blink::CompositedLayerMapping' and its parent.