Revert "Execute plugin notifications synchronously (#32)"
The specific reason this is being reverted is that reading the properties
'bbOnUpdateVisibility' and 'bbOnUpdateGeometry' from the plugin's DOM event
triggers a layout (and maybe style recalculation). During UI initialization,
this could cause an illegal DOM document lifecycle transition that would
trigger an assertion in a debug build and probably just terminated layout
in a release build.

This reverts commit c19bbea0535bd27c2dc785ed40ba884514745f9a.