Allow embedder to indicate failure when creating 'CreateCompositorFrameSink'
Use 'base::Optional' to store the result of
'content::ContentRendererClient::CreateCompositorFrameSink'. This allows the
embedder to indicate that the call failed (by setting the 'optional' to be
a null 'unique_ptr') as well as that the call isn't implemented by the
embedder (by returning an unset 'optional').

This fixes an issue in 'renderer UI' mode where
'blpwtk2::RenderCompositor::CreateCompositorFrameSink()' was getting called for
a 'content::RenderView' before it was associated with a HWND.

In that case, control would pass to the default implementation of
'content::RenderThreadImpl::CreateCompositorFrameSink()' which would end up
presenting black or empty windows.

Now, 'blpwtk2::RenderCompositor::CreateCompositorFrameSink()' can report a
failure and 'content::RenderWidgetCompositor::RequestNewCompositorFrameSink()'
will attempt to re-try the request asynchronously.