Renderer UI: Block renderer during 'DisableSwapUntilResize()'
Calling 'cc::Display::Resize()' with a size of (0, 0) is intended to prevent
the display from swapping incorrectly-sized frames to the window (which can
cause noticeable stretching of the frame with Direct3D).

In 'Renderer UI' mode, this call is executed on the compositor's 'impl' thread.
This change causes the main thread to block until the function returns, which
eliminates the much-hated 'jello' effect during window resizing.