Merge-69 [Autofill] Keeping the user edit.
Problem:
When we Autofill, to decide if we should fill a field or leave it as
is, we check if the last edit was user's edit. If so, we would not
autofill (clear, etc). But sometimes, the website writes on the top
of the autofilled value (exp: formatting a phone number). So, if the
user adds a value on the field, the JS would edit it, and it would be
overwritten by autofill.
Solution:
Instead of looking at the last edit, look if the user has
ever edited that field.

Bug: 860542, 867556
Change-Id: I59ec9c1efab98a9a07f23a47a0354e04ba9b87aa
Reviewed-on: https://chromium-review.googlesource.com/1150683
Reviewed-by: Sebastien Seguin-Gagnon <sebsg@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Parastoo Geranmayeh <parastoog@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#579038}(cherry picked from commit 933a488d53c7c76d966744e327038d3c3e96307a [formerly e9e9833d8b0bedfd9d1a43084d2cd1aed13ef469])
Reviewed-on: https://chromium-review.googlesource.com/1156766
Cr-Commit-Position: refs/branch-heads/3497@{#291}
Cr-Branched-From: eb3b62d2dd0f40372af761b0411c411a1a268eca [formerly 271eaf50594eb818c9295dc78d364aea18c82ea8]-refs/heads/master@{#576753}
Former-commit-id: 2171da0f30ec3b712f4d23363c5950d38a8a140b