[iOS] Fix white flicker in iPad tab grid transitions.
Exiting the tab grid on iPad, especially for tabs close to the top of the screen, shows a white flicker in the toolbar area of the tab.

This is due to the fade in/fade out of the grid cell header and the toolbar snapshot, which transitions through a state where the header is totally white. The timing is such that this can happen right when the top of the tab is in position, making for a distinct flash in the toolbar.

This CL fixes this by using a crossfade instead of the fade in/ fade out. The transition is fast enough that the crossfade doesn't look muddy, and it remove the abrupt flash.

Bug: 865440
Cq-Include-Trybots: luci.chromium.try:ios-simulator-full-configs;master.tryserver.chromium.mac:ios-simulator-cronet
Change-Id: Ifbceef6a79cbde4ea08c76b6f7e6a35bd8225330
Reviewed-on: https://chromium-review.googlesource.com/1158832
Commit-Queue: edchin <edchin@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#579888}(cherry picked from commit 9ab7e46a1224aab98e13ade3c97b060a943c44d9 [formerly 01191ed1f32cbccb798f6df1d7e67ea6c5aa5cdf])
Reviewed-on: https://chromium-review.googlesource.com/1161063
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/branch-heads/3497@{#375}
Cr-Branched-From: eb3b62d2dd0f40372af761b0411c411a1a268eca [formerly 271eaf50594eb818c9295dc78d364aea18c82ea8]-refs/heads/master@{#576753}
Former-commit-id: c25402f83ef9f6135e571d485e463e061e0932de