Merge M69: "Temporarily use null seed DB for OTR VideoDecodePerfHistory"
The VideoDecodePerfHistory is used to make smoothness predictions for
the MediaCapabilities API. For normal profiles, the perf stats are to
a saved to small DB on disk. For incognito profiles we instead save the
stats "in memory" and they disappear when the session ends.

The lines removed in this change "seeded" the Incoginto profile with
the performance stats of the original profile. This has caused some
crashes because we appear to be using the seed DB after its freed (or
perhaps corrupted).

The root cause is not yet known, so the temporary fix is to simply
stop providing a seed DB for incognito. This is already the case for
Guest profiles (no notion of an originating profile) and for any
new on-the-record profile.

The impact will be fairly small, as it only takes ~30 seconds of
playback for enough new stats to accrue such that prediction quality
is unaffected.

In a follow up (after the release) I will revert this
and add some observer code to CHECK when seed_db_ is destroyed before
the InMemoryDB.

TBR=chcunningham@chromium.org

(cherry picked from commit 19bdfdb8be52fc49422bf08df4fe78c5f07b4eac [formerly a88a542a9e9a953c4add08272a5c2b40772511f2])

Bug: 865321
Change-Id: I103b4679bcfa99310d44f4bb810db0535f2d44f0
Reviewed-on: https://chromium-review.googlesource.com/1167514
Commit-Queue: Chrome Cunningham <chcunningham@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#581633}
Reviewed-on: https://chromium-review.googlesource.com/1173330
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/branch-heads/3497@{#584}
Cr-Branched-From: eb3b62d2dd0f40372af761b0411c411a1a268eca [formerly 271eaf50594eb818c9295dc78d364aea18c82ea8]-refs/heads/master@{#576753}
Former-commit-id: 11d53ca09388a5cb3d0dde781aa90838e99d9303