[Settings] UserEvents depends on history and passphrase
This is a reland of:
https://chromium-review.googlesource.com/1170202

which was reverted because of missing svg icon.

The user events UI toggle should be grayed out and turned off
when there is a passphrase or when history sync is disabled.
Note that this is purely a UI change: the internal sync datatype
is not really turned off in that case, and we assume that the Sync
engines will correctly stop syncing user events in these cases,
even if the preference is still ON.

Bug: 865522, 865537
Change-Id: I745f048b3ca301231f57c04b82d0d9d82a6f3b90
Reviewed-on: https://chromium-review.googlesource.com/1170825
Reviewed-by: Scott Chen <scottchen@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#582433}(cherry picked from commit 2ece740555c02bf353a9bb84907cbed858c56b8d [formerly 31322898bc81e2e07e3d9f89aa051a60c5f93215])
Reviewed-on: https://chromium-review.googlesource.com/1174491
Reviewed-by: Thomas Tangl <tangltom@chromium.org>
Cr-Commit-Position: refs/branch-heads/3497@{#614}
Cr-Branched-From: eb3b62d2dd0f40372af761b0411c411a1a268eca [formerly 271eaf50594eb818c9295dc78d364aea18c82ea8]-refs/heads/master@{#576753}
Former-commit-id: 5cc3e5c6271d6110c632a60323ba4e208b5e78d6