Don't assume clipboard text is null terminated
Prior to this commit, ClipboardUtil::GetPlainText copied the clipboard
text onto the output parameter without explicitly specifying the string.
This caused the copy operation to guess the size by looking for the
null-terminator.  Since the clipboard text isn't guaranteed to be a
c-string, this can lead to the copy operating reading more than it needs
to.  This commit makes sure that the string size is explicitly specified
so the correct number of characters are copied.