Disable UpdateHoverPostLayout
This new Webkit feature tries to fire a fake mouse event at post layout stage to
update the hover state. However, this feature is buggy. During resizing the window, the
mouse events leaks to the DOM and some divs got hover effect.
https://bugs.chromium.org/p/chromium/issues/detail?id=811041&can=1&q=resize%20mouseover&colspec=ID%20Pri%20M%20Stars%20ReleaseBlock%20Component%20Status%20Owner%20Summary%20OS%20Modified
https://bugs.chromium.org/p/chromium/issues/detail?id=811041&can=1&q=resize%20mouseover&sort=-modified&colspec=ID%20Pri%20M%20Stars%20ReleaseBlock%20Component%20Status%20Owner%20Summary%20OS%20Modified

To repro this in R+,
1. Run {TKS //SBOX}
2. When resizing the window, some table cells got highlighted(:hover).

Similar issue is also noticed in BEAM app.
This feature seems to be solid in CR67, so we'll enable it back in CR67