Introduce 'blpwtk2::WebScene'
Adds a new implementation of 'blpwtk2::WebView', called 'blpwtk2::WebScene'.
The purpose of this implementation is to synchronously provide a V8 context
with the DOM's globals as well as a DOM populated with some optionally
specified HTML. This webview does not create any browser-side resources, and
cannot be displayed or interacted with, although it does implement the
'drawContentsToBlob()' and 'handleInputEvents()' functions, so could be used
to rasterize the DOM to an image and to simulate user interaction.

Calling 'blpwtk2::Toolkit::createWebScene()' will return a new instance of this
class.