Remember a successful upload of an empty EID for the session
If a new instance of the class is created (as a result of a new UI
session/Chromium being run) we will forget that we did upload. This
is fine, as the behavior we want is to not repeatedly respond to a
server's requests for upload if we already uploaded (due to a TPM
bug that could cause this behavior) while still being able to upload
later after a powerwash or an OS update and reboot.

BUG=chromium:867724
TEST=unit_tests --gtest_filter=*EnrollmentPolicyObserver*

Change-Id: Ic9b080e67aa37a97148bffc04b0da36c9b040a99
Reviewed-on: https://chromium-review.googlesource.com/1213267
Reviewed-by: James Cook <jamescook@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Yves Arrouye <drcrash@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#590770}(cherry picked from commit 067aadbdf2768d7a57a70a0f06eda64dc4647466 [formerly ba619c1cad12e74a133425e2b062fb36ab856035])
Reviewed-on: https://chromium-review.googlesource.com/1222279
Reviewed-by: Yves Arrouye <drcrash@chromium.org>
Cr-Commit-Position: refs/branch-heads/3497@{#930}
Cr-Branched-From: eb3b62d2dd0f40372af761b0411c411a1a268eca [formerly 271eaf50594eb818c9295dc78d364aea18c82ea8]-refs/heads/master@{#576753}
Former-commit-id: fe9c39ad3d8523cf60cce567b122d554ce1f5ea6