Use LayoutRect internally not to lose decimal points.
WebViewImpl::ExpandRubberbandRect() was using LayoutRect internally to
calculate the rubberband area, but returning as WebRect which only holds
integer values, resulting in lost decimals. This was causing a problem when
FinishRubberBanding() got this WebRect value to grab texts as more texts were
captured than the area. The change here is to use LayoutRect in all related
functions to keep the exact area. The publicly exposed APIs are still use
WebRect.