Remove 'std::function' from 'v8::StartupData'
'v8::StartupData' was given a 'std::function' to ensure that memory was
being freed in the same dll as where the memory was allocated.

Presumably due to differences in the layout of 'std::function' between
VS2013 and VS2015, the presence of this 'std::function' caused VS2013 to
create incorrect code, corrupting nearby values, ultimately leading to
test failures in 'ap8'.

The function was carrying no state, so it can be replaced with a plain
function pointer, which is safe to use across dll boundaries, and causes
the correct code to be generated.