Add support for text-align in SetText optimization
This patch adds support for center and right/end text alignments,
as we know for sure we're in left-to-right
the operation is pretty simple.
Justify alignment is still skipped as we need to layout the text
anyway in that case.

Bug: 805785

Backport of a patch by Manuel Rego Casasnovas <rego@igalia.com>