Check if the object is LayoutListItem before casting
This is to fix {DRQS 132841802}. When indenting the first <LI> in a <OL>, a new <OL> seems to be added
as its sibling, then it moves under the new <OL> later. It makes the code to
find the new <OL> and incorrectly cast it to LayoutListItem. This change checks if
it is a LayoutListItem.

Also changed to return LayoutUnit() to sync the style with the rest of the
code