Web Script Context: Fix Context Check for Workers
It is seemingly invalid to call
'WrapperCreationSecurityCheck::VerifyContextAccess' with a web worker
context.  To avoid crashing when web workers run, we only call this
method when the currently installed context doesn't match the creation
context.  In web workers there is only ever a single context (as far as
I can tell), so the call is not made.