Fix crash on SetText optimization
This optimization was introduced in r578233, but with the test included in this
patch it was causing a crash. The problem is that the RootInlineBox has more
than one children which is not expected. In that case we shouldn't apply the
optimization, so a new condition has been added in
LayoutText::CanOptimizeSetText() to avoid it.

Backport of a patch by Manuel Rego Casasnovas <rego@igalia.com>