[css-contain] Fix wrapping issue in SetText() optimization
The optimization introduced in r578233 was not taking into account
one case. If we have different LayoutText under the same parent object
we cannot apply the optimization as we need to recompute wrapping.

To fix that we need to check that the RootInlineBox has no siblings.

Backport of a patch by Manuel Rego Casasnovas <rego@igalia.com>