Renderer UI: Compatibility with DirectComposition
When DirectComposition is used, the GPU process hosts an HWND that displays
the compositor's output; this HWND is made a 'child' of the HWND created for
the webview.

Since the webview HWND is ordinarily owned by the browser process, it is
responsible for establishing the relationship with the compositor HWND by
using a class called 'ui::gfx::RenderingWindowManager' in response to an
IPC message from the GPU process, 'SetChildSurface()'.

When 'renderer UI' mode is enabled, the webview HWND is owned by a
renderer process. This change adds IPC messages that the renderer can send
to the browser process to make a renderer-owned HWND available to host
compositor output. 'blpwtk2::RenderCompositor' sends these messages via
the 'blpwtk2::ProfileImpl' object.

This commit also makes a small change to
'content::GpuProcessHost::SetChildSurface()' to relax the requirement that
the parent (hosting) HWND must be owned by the browser process.