Introduce the function 'blpwtk2::setSecurityToken()'
This function sets an optional V8 security token on the webview. After the
token is set, it will be applied to the V8 context created for every
subsequent document load in every webframe in the webview.

Calling this function will not affect the security token for webframes that
have already finished loading.