MultiHeapTracer: Set EmbedderHeapTracer::isolate_
'v8::EmbedderHeapTracer' has a member 'isolate_' which is protected.
When 'v8::Isolate::SetEmbedderHeapTracer' is called,
'v8::internal::LocalEmbedderHeapTracer' -- which is declared a friend --
will set 'isolate_' on the 'EmbedderHeapTracer' instance, indicating
that it is now connected to an 'Isolate'.

With the introduction of the 'gin::MultiHeapTracer', an instance of this
class is the only tracer set on the 'Isolate'.  When additional
'EmbedderHeapTracer' objects are set on the 'MultiHeapTracer', we need
to also indicate that they are now connected to an 'Isolate'.  To
achieve this, we declare 'gin::MultiHeapTracer' as a friend of
'v8::EmbedderHeapTracer' so that it can set 'isolate_' accordingly.

We could have alternatively exposed public methods on
'v8::EmbedderHeapTracer' to set 'isolate_', but this would polute the
public interface, so long distance friendship seems like the lesser of
two evils.