Take snapshot without issuing a relayout
The embedder is responsible for preparing a webview or a webframe with
the printable content and waiting for a layout cycle to finish before
it calls blpwtk2::WebView::drawContentsToBlob.  We don't need to issue
a relayout in here.  In fact, issuing a relayout prevents us from
returning an image immediately, since we would have to wait for the
layout to finish.

(cherry picked from commit 289233dcffd0bdc38f33411dd7a9d37d9af1ee20)