Disable resize optimization (#289)
Traditionally, the entry point of UI events for a Chromium browser is the
'browser' process. When the user resizes the browser window, the 'browser'
process sends a message to the 'renderer' process to update the inner content
size. During the update of the inner content size, the compositor within the
'renderer' process prepares OpenGL commands to be sent to the 'gpu' process.
The 'gpu' process checks the size of the OS window owned by the
'browser' process and creates a Direct3D surface of that size. It then
translates the OpenGL commands to Direct3D commands before executing
them on this surface.

To summarize, the event pipeline is:
  browser -> renderer -> gpu

When the 'gpu' process checks the size of the OS window owned by the 'browser',
it can be sure that it's getting the correct size since the resize event first
hits the 'browser' process.

The public API of blpwtk2 allows the embedder to set the size of the webview
from within the 'renderer' process. The 'renderer' begins to resize the inner
content and it also sends a message to the 'browser' process to resize the
OS window. Once the 'renderer' process finishes the inner content resize, its
compositor sends the OpenGL commands to the 'gpu' process. The 'gpu' process
queries the size of the OS window owned by the 'browser' process to determine
the size of the Direct3D surface it needs to create.

To summarize this model, the event pipeline is:
  renderer -> gpu
      \-> browser -> renderer

Since the 'browser' and 'gpu' process gets the size information from the
'renderer' simultanously, it's sometimes possible that the 'gpu' process
queries the OS window (that is owned by the browser) too early, even before
the 'browser' gets the message to resize the window. This doesn't happen very
often because the 'renderer' has to do a lot of compute-expensive work before
sending a message to the 'gpu' process, at which point the 'browser' already
resizes the OS window.

In recent version of Chromium, the probability that the 'gpu' process
queries the size of the OS window too early seems to have increased. I
believe this is due to slimming-paint reducing the amount of work the
compositor does when there is not a lot of changes in the content and so the
'gpu' process begins to process the frame a bit early.