Renderer UI
Introduces a new implementation of 'blpwtk2::WebView', 'blpwtk2::RenderWebView'.
This implementation runs the 'user interface' portion of the webview -
compositing and input event handling - within the renderer thread, bypassing
the browser thread for those tasks. The implementation otherwise delegates
itself to an instance of 'blpwtk2::WebViewProxy'.

To perform compositing in the renderer thread, a new implementation of the
'content::FrameSinkProvider' is provided and bound to the Mojo endpoint that
'content::RenderThreadImpl' uses to create instances of the
'viz::CompositorFrameSink' interface for each 'content::RenderWidget'.

To handle input events, 'blwptk2::RenderWebView' creates a HWND to accept
user input and route it to an instance of 'content::InputRouterImpl'.
That instance is also bound to the Mojo endpoint that the renderer uses
to send responses back to the browser.

Passing 'true' to 'blpwtk2::ToolkitCreateParams::setRendererUIEnabled()'
when initializing the toolkit will cause a 'blpwtk2::RenderWebView' to be
returned for all requested web views.