Fix crash on SetText optimization
This optimization was introduced in r578233
but it's causing a crash when you use "text-overflow: ellipsis".
We need to check if we need or not an ellipsis in the new text,
and compute the position of the ellipsis if it's required.
For that reason we cannot apply the optimization
and we need to do a proper layout.
The patch adds a new condition in LayoutText::CanOptimizeSetText()
to avoid applying the optimization in that case.

BUG=805785

Backport of a patch by Manuel Rego Casasnovas <rego@igalia.com>