Remove an unnecessary call to a string conversion function
The input string is already UTF16.

This change is necessary because the existing code interacted poorly
with changes made in SHA 8244f04d7167a596d4d6815df98542b478ea227d.

The commit logs indicate that the erroneous call was part of changes related
to 'docprinter', but the call was not actually in the 'bb70/blpwtk2/docprinter'
branch.

So, this fix is being merged directly into the release branch.