Fix Crash When Resuming From Breakpoints (#351)
When a JavaScript breakpoint is triggered, Chromium runs an inner
message loop.  When this message loop finishes running, Chromium has an
assumption that there is an event queue running on the main thread.
When running 'blpwtk2_shell.exe' and stopping at breakpoints, this is
the case, as all JavaScript is run from the 'MainMessagePump'.  When
Chromium is embedded such that JavaScript can run outside of the
'MainMessagePump', however, there is no running event queue.  This
breaks Chromium's assumption, leading to a crash.

This commit removes this assumption, allowing the code to work
regardless of whether a breakpoint was triggered during JavaScript
running as part of a Chromium event, or not.